requirements.txt
az-iranian-bank-gateways==1.6.18
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.3
django-apscheduler==0.6.2
django-heroku==0.3.1
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pycoin==0.92.20220529
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.2
urllib3==1.26.9
whitenoise==6.2.0